requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
Pillow==10.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
PyJWT==2.8.0
cryptography==41.0.7